+++ Runtime/Collections/GridMap.cs
@@ -1,6 +1,4 @@
-﻿using System.Security.Cryptography.X509Certificates;
-using Unity.Mathematics;
-using UnityEngine;
+﻿using Unity.Mathematics;
 using UnityEngine.Assertions;
 
 namespace Timespawn.Core.Collections